@@ -1,23 +1,22 @@
-﻿namespace vue_expenses_api.Dtos
+﻿namespace vue_expenses_api.Dtos;
+
+public class ExpenseTypeDto 
 {
-    public class ExpenseTypeDto 
+    protected ExpenseTypeDto()
     {
-        protected ExpenseTypeDto()
-        {
-        }
-
-        public ExpenseTypeDto(
-            int id,
-            string name,
-            string description)
-        {
-            Id = id;
-            Name = name;
-            Description = description;
-        }
-
-        public int Id { get; set; }
-        public string Name { get; set; }
-        public string Description { get; set; }
     }
+
+    public ExpenseTypeDto(
+        int id,
+        string name,
+        string description)
+    {
+        Id = id;
+        Name = name;
+        Description = description;
+    }
+
+    public int Id { get; set; }
+    public string Name { get; set; }
+    public string Description { get; set; }
 }
