@@ -6,56 +6,55 @@ using MediatR;
 using Microsoft.EntityFrameworkCore;
 using vue_expenses_api.Infrastructure;
 
-namespace vue_expenses_api.Features.ExpenseCategories
+namespace vue_expenses_api.Features.ExpenseCategories;
+
+public class DeleteExpenseCategory
 {
-    public class DeleteExpenseCategory
+    public class Command : IRequest
     {
-        public class Command : IRequest
+        public Command(
+            int id)
         {
-            public Command(
-                int id)
-            {
-                Id = id;
-            }
-
-            public int Id { get; set; }
+            Id = id;
         }
 
-        public class CommandValidator : AbstractValidator<DeleteExpenseCategory.Command>
+        public int Id { get; set; }
+    }
+
+    public class CommandValidator : AbstractValidator<DeleteExpenseCategory.Command>
+    {
+        public CommandValidator()
         {
-            public CommandValidator()
-            {
-                RuleFor(x => x.Id).NotNull().NotEmpty();
-            }
+            RuleFor(x => x.Id).NotNull().NotEmpty();
+        }
+    }
+
+    public class Handler : IRequestHandler<Command, Unit>
+    {
+        private readonly ExpensesContext _context;
+
+        public Handler(
+            ExpensesContext db)
+        {
+            _context = db;
         }
 
-        public class Handler : IRequestHandler<Command, Unit>
+        public async Task<Unit> Handle(
+            Command request,
+            CancellationToken cancellationToken)
         {
-            private readonly ExpensesContext _context;
+            var expenseCategory = await _context.ExpenseCategories.FirstOrDefaultAsync(
+                x => x.Id == request.Id,
+                cancellationToken);
 
-            public Handler(
-                ExpensesContext db)
+            if (expenseCategory == null)
             {
-                _context = db;
+                throw new Exception("Not Found");
             }
 
-            public async Task<Unit> Handle(
-                Command request,
-                CancellationToken cancellationToken)
-            {
-                var expenseCategory = await _context.ExpenseCategories.FirstOrDefaultAsync(
-                    x => x.Id == request.Id,
-                    cancellationToken);
-
-                if (expenseCategory == null)
-                {
-                    throw new Exception("Not Found");
-                }
-
-                expenseCategory.Archive();
-                await _context.SaveChangesAsync(cancellationToken);
-                return Unit.Value;
-            }
+            expenseCategory.Archive();
+            await _context.SaveChangesAsync(cancellationToken);
+            return Unit.Value;
         }
     }
 }
