@@ -8,32 +8,32 @@ using MediatR;
 using vue_expenses_api.Dtos;
 using vue_expenses_api.Infrastructure.Security;
 
-namespace vue_expenses_api.Features.Statistics
+namespace vue_expenses_api.Features.Statistics;
+
+public class DashboardCategoryStatisticsList
 {
-    public class DashboardCategoryStatisticsList
+    public class Query : IRequest<List<CategoryStatisticsDto>>
     {
-        public class Query : IRequest<List<CategoryStatisticsDto>>
+    }
+
+    public class QueryHandler : IRequestHandler<Query, List<CategoryStatisticsDto>>
+    {
+        private readonly IDbConnection _dbConnection;
+        private readonly ICurrentUser _currentUser;
+
+        public QueryHandler(
+            IDbConnection dbConnection,
+            ICurrentUser currentUser)
         {
+            _dbConnection = dbConnection;
+            _currentUser = currentUser;
         }
 
-        public class QueryHandler : IRequestHandler<Query, List<CategoryStatisticsDto>>
+        public async Task<List<CategoryStatisticsDto>> Handle(
+            Query request,
+            CancellationToken cancellationToken)
         {
-            private readonly IDbConnection _dbConnection;
-            private readonly ICurrentUser _currentUser;
-
-            public QueryHandler(
-                IDbConnection dbConnection,
-                ICurrentUser currentUser)
-            {
-                _dbConnection = dbConnection;
-                _currentUser = currentUser;
-            }
-
-            public async Task<List<CategoryStatisticsDto>> Handle(
-                Query request,
-                CancellationToken cancellationToken)
-            {
-                var sql = $@"WITH RECURSIVE Months(month) AS (
+            var sql = $@"WITH RECURSIVE Months(month) AS (
 								SELECT 
 									strftime('%m', 'now')
 								UNION ALL
@@ -72,15 +72,14 @@ namespace vue_expenses_api.Features.Statistics
 								Month,
 								ec.Name";
 
-                var expenses = await _dbConnection.QueryAsync<CategoryStatisticsDto>(
-                    sql,
-                    new
-                    {
-                        userEmailId = _currentUser.EmailId
-                    });
+            var expenses = await _dbConnection.QueryAsync<CategoryStatisticsDto>(
+                sql,
+                new
+                {
+                    userEmailId = _currentUser.EmailId
+                });
 
-                return expenses.ToList();
-            }
+            return expenses.ToList();
         }
     }
 }
