@@ -10,92 +10,91 @@ using vue_expenses_api.Infrastructure;
 using vue_expenses_api.Infrastructure.Exceptions;
 using vue_expenses_api.Infrastructure.Security;
 
-namespace vue_expenses_api.Features.Users
+namespace vue_expenses_api.Features.Users;
+
+public class Login
 {
-    public class Login
+    public class Command : IRequest<UserDto>
     {
-        public class Command : IRequest<UserDto>
+        public Command(
+            string email,
+            string password)
         {
-            public Command(
-                string email,
-                string password)
-            {
-                Email = email;
-                Password = password;
-            }
-
-            public string Email { get; set; }
-            public string Password { get; set; }
-
+            Email = email;
+            Password = password;
         }
 
-        public class CommandValidator : AbstractValidator<Command>
+        public string Email { get; set; }
+        public string Password { get; set; }
+
+    }
+
+    public class CommandValidator : AbstractValidator<Command>
+    {
+        public CommandValidator()
         {
-            public CommandValidator()
-            {
-                RuleFor(x => x.Email).NotNull().NotEmpty();
-                RuleFor(x => x.Password).NotNull().NotEmpty();
-            }
+            RuleFor(x => x.Email).NotNull().NotEmpty();
+            RuleFor(x => x.Password).NotNull().NotEmpty();
+        }
+    }
+
+    public class Handler : IRequestHandler<Command, UserDto>
+    {
+        private readonly IPasswordHasher _passwordHasher;
+        private readonly IJwtTokenGenerator _jwtTokenGenerator;
+        private readonly ExpensesContext _context;
+
+        public Handler(
+            IPasswordHasher passwordHasher,
+            IJwtTokenGenerator jwtTokenGenerator,
+            ExpensesContext context)
+        {
+            _passwordHasher = passwordHasher;
+            _jwtTokenGenerator = jwtTokenGenerator;
+            _context = context;
         }
 
-        public class Handler : IRequestHandler<Command, UserDto>
+        public async Task<UserDto> Handle(
+            Command request,
+            CancellationToken cancellationToken)
         {
-            private readonly IPasswordHasher _passwordHasher;
-            private readonly IJwtTokenGenerator _jwtTokenGenerator;
-            private readonly ExpensesContext _context;
+            var user = await _context.Users.SingleOrDefaultAsync(
+                x => x.Email == request.Email && !x.Archived,
+                cancellationToken);
 
-            public Handler(
-                IPasswordHasher passwordHasher,
-                IJwtTokenGenerator jwtTokenGenerator,
-                ExpensesContext context)
+            if (user == null)
             {
-                _passwordHasher = passwordHasher;
-                _jwtTokenGenerator = jwtTokenGenerator;
-                _context = context;
+                throw new HttpException(
+                    HttpStatusCode.Unauthorized,
+                    new {Error = "Invalid credentials."});
             }
 
-            public async Task<UserDto> Handle(
-                Command request,
-                CancellationToken cancellationToken)
-            {
-                var user = await _context.Users.SingleOrDefaultAsync(
-                    x => x.Email == request.Email && !x.Archived,
-                    cancellationToken);
-
-                if (user == null)
-                {
-                    throw new HttpException(
-                        HttpStatusCode.Unauthorized,
-                        new {Error = "Invalid credentials."});
-                }
-
-                if (!user.Hash.SequenceEqual(
+            if (!user.Hash.SequenceEqual(
                     _passwordHasher.Hash(
                         request.Password,
                         user.Salt)))
-                {
-                    throw new HttpException(
-                        HttpStatusCode.Unauthorized,
-                        new {Error = "Invalid credentials."});
-                }
-                // generate refresh token
-                var refreshToken = _jwtTokenGenerator.GenerateRefreshToken();
-                user.AddRefreshToken(refreshToken, user.Id);
-                
-                var token = await _jwtTokenGenerator.CreateToken(request.Email);
-                await _context.SaveChangesAsync(cancellationToken);
-                
-                return new UserDto(
-                    user.FirstName,
-                    user.LastName,
-                    user.FullName,
-                    user.SystemName,
-                    user.Email,
-                    token,
-                    refreshToken,
-                    user.CurrencyRegionName,
-                    user.UseDarkMode);
+            {
+                throw new HttpException(
+                    HttpStatusCode.Unauthorized,
+                    new {Error = "Invalid credentials."});
             }
+            // generate refresh token
+            var refreshToken = _jwtTokenGenerator.GenerateRefreshToken();
+            user.AddRefreshToken(refreshToken, user.Id);
+                
+            var token = await _jwtTokenGenerator.CreateToken(request.Email);
+            await _context.SaveChangesAsync(cancellationToken);
+                
+            return new UserDto(
+                user.FirstName,
+                user.LastName,
+                user.FullName,
+                user.SystemName,
+                user.Email,
+                token,
+                refreshToken,
+                user.CurrencyRegionName,
+                user.UseDarkMode);
         }
     }
 }
