@@ -7,64 +7,63 @@ using vue_expenses_api.Dtos;
 using vue_expenses_api.Infrastructure;
 using vue_expenses_api.Infrastructure.Security;
 
-namespace vue_expenses_api.Features.Users
+namespace vue_expenses_api.Features.Users;
+
+public class UpdateProfile
 {
-    public class UpdateProfile
+    public class Command : IRequest<ProfileDetailsDto>
     {
-        public class Command : IRequest<ProfileDetailsDto>
+        public Command(
+            string firstName,
+            string lastName)
         {
-            public Command(
-                string firstName,
-                string lastName)
-            {
-                FirstName = firstName;
-                LastName = lastName;
-            }
-
-            public string FirstName { get; set; }
-            public string LastName { get; set; }
+            FirstName = firstName;
+            LastName = lastName;
         }
 
-        public class CommandValidator : AbstractValidator<Command>
+        public string FirstName { get; set; }
+        public string LastName { get; set; }
+    }
+
+    public class CommandValidator : AbstractValidator<Command>
+    {
+        public CommandValidator()
         {
-            public CommandValidator()
-            {
-                RuleFor(x => x.FirstName).NotEmpty().NotNull();
-            }
+            RuleFor(x => x.FirstName).NotEmpty().NotNull();
+        }
+    }
+
+    public class Handler : IRequestHandler<Command, ProfileDetailsDto>
+    {
+        private readonly ExpensesContext _context;
+        private readonly ICurrentUser _currentUser;
+
+        public Handler(
+            ExpensesContext db,
+            ICurrentUser currentUser)
+        {
+            _context = db;
+            _currentUser = currentUser;
         }
 
-        public class Handler : IRequestHandler<Command, ProfileDetailsDto>
+        public async Task<ProfileDetailsDto> Handle(
+            Command request,
+            CancellationToken cancellationToken)
         {
-            private readonly ExpensesContext _context;
-            private readonly ICurrentUser _currentUser;
+            var user = await _context.Users.SingleAsync(
+                x => x.Email == _currentUser.EmailId,
+                cancellationToken);
 
-            public Handler(
-                ExpensesContext db,
-                ICurrentUser currentUser)
-            {
-                _context = db;
-                _currentUser = currentUser;
-            }
+            user.UpdateName(
+                request.FirstName,
+                request.LastName);
 
-            public async Task<ProfileDetailsDto> Handle(
-                Command request,
-                CancellationToken cancellationToken)
-            {
-                var user = await _context.Users.SingleAsync(
-                    x => x.Email == _currentUser.EmailId,
-                    cancellationToken);
+            await _context.SaveChangesAsync(cancellationToken);
 
-                user.UpdateName(
-                    request.FirstName,
-                    request.LastName);
-
-                await _context.SaveChangesAsync(cancellationToken);
-
-                return new ProfileDetailsDto(
-                    user.FirstName,
-                    user.LastName,
-                    user.FullName);
-            }
+            return new ProfileDetailsDto(
+                user.FirstName,
+                user.LastName,
+                user.FullName);
         }
     }
 }
