@@ -4,60 +4,59 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Newtonsoft.Json;
 
-namespace vue_expenses_api.Infrastructure.Exceptions
+namespace vue_expenses_api.Infrastructure.Exceptions;
+
+public class ErrorHandlingMiddleware
 {
-    public class ErrorHandlingMiddleware
+    private readonly RequestDelegate _next;
+
+    public ErrorHandlingMiddleware(
+        RequestDelegate next)
     {
-        private readonly RequestDelegate _next;
+        _next = next;
+    }
 
-        public ErrorHandlingMiddleware(
-            RequestDelegate next)
+    public async Task Invoke(
+        HttpContext context)
+    {
+        try
         {
-            _next = next;
+            await _next(context);
+        }
+        catch (Exception ex)
+        {
+            await HandleExceptionAsync(
+                context,
+                ex);
+        }
+    }
+
+    private static async Task HandleExceptionAsync(
+        HttpContext context,
+        Exception exception)
+    {
+        object errors = null;
+
+        switch (exception)
+        {
+            case HttpException re:
+                errors = re.Errors;
+                context.Response.StatusCode = (int) re.Code;
+                break;
+            case Exception e:
+                errors = string.IsNullOrWhiteSpace(e.Message) ? "Error" : e.Message;
+                context.Response.StatusCode = (int) HttpStatusCode.InternalServerError;
+                break;
         }
 
-        public async Task Invoke(
-            HttpContext context)
-        {
-            try
+        context.Response.ContentType = "application/json";
+
+        var result = JsonConvert.SerializeObject(
+            new
             {
-                await _next(context);
-            }
-            catch (Exception ex)
-            {
-                await HandleExceptionAsync(
-                    context,
-                    ex);
-            }
-        }
+                errors
+            });
 
-        private static async Task HandleExceptionAsync(
-            HttpContext context,
-            Exception exception)
-        {
-            object errors = null;
-
-            switch (exception)
-            {
-                case HttpException re:
-                    errors = re.Errors;
-                    context.Response.StatusCode = (int) re.Code;
-                    break;
-                case Exception e:
-                    errors = string.IsNullOrWhiteSpace(e.Message) ? "Error" : e.Message;
-                    context.Response.StatusCode = (int) HttpStatusCode.InternalServerError;
-                    break;
-            }
-
-            context.Response.ContentType = "application/json";
-
-            var result = JsonConvert.SerializeObject(
-                new
-                {
-                    errors
-                });
-
-            await context.Response.WriteAsync(result);
-        }
+        await context.Response.WriteAsync(result);
     }
 }
