@@ -2,17 +2,16 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.ApplicationModels;
 
-namespace vue_expenses_api.Infrastructure
-{
-    public class GroupByApiRootConvention : IControllerModelConvention
-    {
-        public void Apply(
-            ControllerModel controller)
-        {
-            var controllerNamespace = controller.Attributes.OfType<RouteAttribute>().FirstOrDefault();
-            var apiVersion = controllerNamespace?.Template?.Split('/')?.First()?.ToLower() ?? "default";
+namespace vue_expenses_api.Infrastructure;
 
-            controller.ApiExplorer.GroupName = apiVersion;
-        }
+public class GroupByApiRootConvention : IControllerModelConvention
+{
+    public void Apply(
+        ControllerModel controller)
+    {
+        var controllerNamespace = controller.Attributes.OfType<RouteAttribute>().FirstOrDefault();
+        var apiVersion = controllerNamespace?.Template?.Split('/')?.First()?.ToLower() ?? "default";
+
+        controller.ApiExplorer.GroupName = apiVersion;
     }
 }
