@@ -1,7 +1,6 @@
-namespace vue_expenses_api.Infrastructure.Security
+namespace vue_expenses_api.Infrastructure.Security;
+
+public class PasswordHasherSettings
 {
-    public class PasswordHasherSettings
-    {
-        public string Key { get; set; }
-    }
+    public string Key { get; set; }
 }
