@@ -5,38 +5,37 @@ using System.Threading.Tasks;
 using FluentValidation;
 using MediatR;
 
-namespace vue_expenses_api.Infrastructure.Validation
+namespace vue_expenses_api.Infrastructure.Validation;
+
+//https://stackoverflow.com/a/42289038
+public class ValidationPipelineBehavior<TRequest, TResponse>
+    : IPipelineBehavior<TRequest, TResponse> where TRequest : IRequest<TResponse>
 {
-    //https://stackoverflow.com/a/42289038
-    public class ValidationPipelineBehavior<TRequest, TResponse>
-        : IPipelineBehavior<TRequest, TResponse> where TRequest : IRequest<TResponse>
+    private readonly List<IValidator<TRequest>> _validators;
+
+    public ValidationPipelineBehavior(
+        IEnumerable<IValidator<TRequest>> validators)
     {
-        private readonly List<IValidator<TRequest>> _validators;
+        _validators = validators.ToList();
+    }
 
-        public ValidationPipelineBehavior(
-            IEnumerable<IValidator<TRequest>> validators)
+    public async Task<TResponse> Handle(
+        TRequest request,
+        RequestHandlerDelegate<TResponse> next,
+        CancellationToken cancellationToken)
+    {
+        var context = new ValidationContext<TRequest>(request);
+        var failures = _validators
+            .Select(v => v.Validate(context))
+            .SelectMany(result => result.Errors)
+            .Where(f => f != null)
+            .ToList();
+
+        if (failures.Count != 0)
         {
-            _validators = validators.ToList();
+            throw new ValidationException(failures);
         }
 
-        public async Task<TResponse> Handle(
-            TRequest request,
-            CancellationToken cancellationToken,
-            RequestHandlerDelegate<TResponse> next)
-        {
-            var context = new ValidationContext<TRequest>(request);
-            var failures = _validators
-                .Select(v => v.Validate(context))
-                .SelectMany(result => result.Errors)
-                .Where(f => f != null)
-                .ToList();
-
-            if (failures.Count != 0)
-            {
-                throw new ValidationException(failures);
-            }
-
-            return await next();
-        }
+        return await next();
     }
 }
